requirements.txt
ag2==0.8.5
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.8.5
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
config==0.5.1
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
frozendict==2.4.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langsmith==0.3.27
lxml==5.3.2
marshmallow==3.26.1
multidict==6.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
propcache==0.3.1
pyautogen==0.8.5
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytils==0.4.3
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
termcolor==3.0.1
tiktoken==0.9.0
tools==0.1.9
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
yfinance==0.2.55
zstandard==0.23.0
